% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispatch.r
\name{bv_render}
\alias{bv_render}
\title{Render Variable Tables/Visualizations by Role}
\usage{
bv_render(x, form)
}
\arguments{
\item{x}{a data.frame object.}

\item{form}{a formula containing the role relationships.}
}
\description{
Render Variable Tables/Visualizations by Role
}
